Add doc examples to ~50 most-used public methods
Add # Examples sections with runnable doc tests to the most important
public APIs: Command (9), Event (10), RuntimeConfig (6), Runtime (8),
TestHarness (5), AppHarness (4), EventQueue (4), Theme (3).

Also extract RuntimeConfig to config.rs and terminal mode to terminal.rs
to keep runtime/mod.rs under the 1000-line limit (854 lines).

Doc test count: 440 -> 488 (+48 new runnable examples).
