+++ src/harness/app_harness/mod.rs
@@ -55,6 +55,28 @@ impl<A: App> AppHarness<A> {
     /// Creates a new async test harness with the given dimensions.
     ///
     /// Note: For time control, use `#[tokio::test(start_paused = true)]`.
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// # use envision::prelude::*;
+    /// # struct MyApp;
+    /// # #[derive(Default, Clone)]
+    /// # struct MyState;
+    /// # #[derive(Clone)]
+    /// # enum MyMsg {}
+    /// # impl App for MyApp {
+    /// #     type State = MyState;
+    /// #     type Message = MyMsg;
+    /// #     fn init() -> (MyState, Command<MyMsg>) { (MyState, Command::none()) }
+    /// #     fn update(state: &mut MyState, msg: MyMsg) -> Command<MyMsg> { Command::none() }
+    /// #     fn view(state: &MyState, frame: &mut Frame) {}
+    /// # }
+    /// use envision::harness::AppHarness;
+    ///
+    /// let harness = AppHarness::<MyApp>::new(80, 24)?;
+    /// # Ok::<(), std::io::Error>(())
+    /// ```
     pub fn new(width: u16, height: u16) -> io::Result<Self> {
         let runtime = Runtime::virtual_terminal(width, height)?;
         Ok(Self { runtime })
@@ -144,6 +166,33 @@ impl<A: App> AppHarness<A> {
     ///
     /// This dispatches the message, spawns any async commands, and processes
     /// any immediately available async results.
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// # use envision::prelude::*;
+    /// # struct MyApp;
+    /// # #[derive(Default, Clone)]
+    /// # struct MyState { count: i32 }
+    /// # #[derive(Clone)]
+    /// # enum MyMsg { Increment }
+    /// # impl App for MyApp {
+    /// #     type State = MyState;
+    /// #     type Message = MyMsg;
+    /// #     fn init() -> (MyState, Command<MyMsg>) { (MyState::default(), Command::none()) }
+    /// #     fn update(state: &mut MyState, msg: MyMsg) -> Command<MyMsg> {
+    /// #         match msg { MyMsg::Increment => state.count += 1 }
+    /// #         Command::none()
+    /// #     }
+    /// #     fn view(state: &MyState, frame: &mut Frame) {}
+    /// # }
+    /// use envision::harness::AppHarness;
+    ///
+    /// let mut harness = AppHarness::<MyApp>::new(80, 24)?;
+    /// harness.dispatch(MyMsg::Increment);
+    /// assert_eq!(harness.state().count, 1);
+    /// # Ok::<(), std::io::Error>(())
+    /// ```
     pub fn dispatch(&mut self, msg: A::Message) {
         self.runtime.dispatch(msg);
         self.runtime.process_pending();
@@ -229,6 +278,30 @@ impl<A: App> AppHarness<A> {
     }
 
     /// Runs a single tick of the application.
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// # use envision::prelude::*;
+    /// # struct MyApp;
+    /// # #[derive(Default, Clone)]
+    /// # struct MyState;
+    /// # #[derive(Clone)]
+    /// # enum MyMsg {}
+    /// # impl App for MyApp {
+    /// #     type State = MyState;
+    /// #     type Message = MyMsg;
+    /// #     fn init() -> (MyState, Command<MyMsg>) { (MyState, Command::none()) }
+    /// #     fn update(state: &mut MyState, msg: MyMsg) -> Command<MyMsg> { Command::none() }
+    /// #     fn view(state: &MyState, frame: &mut Frame) {}
+    /// # }
+    /// use envision::harness::AppHarness;
+    ///
+    /// let mut harness = AppHarness::<MyApp>::new(80, 24)?;
+    /// harness.push_event(Event::key(KeyCode::Enter));
+    /// harness.tick()?;
+    /// # Ok::<(), std::io::Error>(())
+    /// ```
     pub fn tick(&mut self) -> io::Result<()> {
         self.runtime.tick()
     }
@@ -286,6 +359,32 @@ impl<A: App> AppHarness<A> {
     /// # Panics
     ///
     /// Panics if the text is not found.
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// # use envision::prelude::*;
+    /// # struct MyApp;
+    /// # #[derive(Default, Clone)]
+    /// # struct MyState;
+    /// # #[derive(Clone)]
+    /// # enum MyMsg {}
+    /// # impl App for MyApp {
+    /// #     type State = MyState;
+    /// #     type Message = MyMsg;
+    /// #     fn init() -> (MyState, Command<MyMsg>) { (MyState, Command::none()) }
+    /// #     fn update(state: &mut MyState, msg: MyMsg) -> Command<MyMsg> { Command::none() }
+    /// #     fn view(state: &MyState, frame: &mut Frame) {
+    /// #         frame.render_widget(ratatui::widgets::Paragraph::new("Welcome"), frame.area());
+    /// #     }
+    /// # }
+    /// use envision::harness::AppHarness;
+    ///
+    /// let mut harness = AppHarness::<MyApp>::new(80, 24)?;
+    /// harness.tick()?;
+    /// harness.assert_contains("Welcome");
+    /// # Ok::<(), std::io::Error>(())
+    /// ```
     pub fn assert_contains(&self, needle: &str) {
         if !self.contains_text(needle) {
             panic!(
